Implement bandit to test security
"bandit" is a Python package that examines code for potential security issues:
hardcoded passwords, SQL injection vectors, and so forth.

The good news is that bandit hasn't found any issues (see the config file for
an explanation of the issues we ignore).  Consequently, there's no changes
needed to the code, and therefore this isn't getting bundled into a version or
a release.  But users may rest easier knowing that we're now actively checking
changes to this code for security issues.